services/ai-service/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
openai==1.17.0
google-generativeai>=0.8.6
